% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Incytr_class.R
\docType{class}
\name{AnyFactor-class}
\alias{AnyFactor-class}
\title{The Incytr Class}
\description{
The Incytr object is created based on the following inputs: single-cell transcriptomic data matrix (normalized), meta data with cell cluster and condition labels.
When inputting an data matrix, genes should be in rows and cells in columns. rownames and colnames should be included.
}
