% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{barcodes_bycondition}
\alias{barcodes_bycondition}
\title{Gather cell barcodes for each condition from the meta data}
\usage{
barcodes_bycondition(object)
}
\arguments{
\item{object}{Incytr object}
}
\value{
a list
}
\description{
Gather cell barcodes for each condition from the meta data
}
