% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinases.R
\name{Cal_EI}
\alias{Cal_EI}
\title{Calculate the exclusiveness index (EI) for the genes between clusters based on the scRNA-seq data}
\usage{
Cal_EI(df, cell_group, fold_threshold = 10)
}
\arguments{
\item{df}{data frame of the expression level in clusters}

\item{cell_group}{the group/cluster to calculate the EI}

\item{fold_threshold}{if the difference between the selected value and the second highest value passes the threshold, then we think it is "highly exclusive", EI = 1. The default setting is fold_threshold = 10.}
}
\value{
a data frame
}
\description{
Calculate the exclusiveness index (EI) for the genes between clusters based on the scRNA-seq data
}
