% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{Cal_foldchange}
\alias{Cal_foldchange}
\title{Calculate the log2FoldChang & adjlog2FC for the data}
\usage{
Cal_foldchange(df, correction = 1e-04, q = NULL)
}
\arguments{
\item{df}{a data frame with the following variables: "gene_symbol", "condition1", "condition2"}

\item{correction}{the value added to all the data to avoid division over zero, the default value is 0.0001}

\item{q}{the percentile used to determine "low expression level" and is used when calculating the adjlog2FC, the default value is 0.75}
}
\value{
a data frame
}
\description{
Calculate the log2FoldChang & adjlog2FC for the data
}
