% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{Cal_PDS}
\alias{Cal_PDS}
\title{Based on the multi-model score, add the KPDS (if applicable) and calculate the PDS (the final score).}
\usage{
Cal_PDS(object, KPDS.weight = NULL, cutoff_PDS = NULL)
}
\arguments{
\item{object}{Incytr object}

\item{KPDS.weight}{weight for the KPDS, the default value is 0.5.}

\item{cutoff_PDS}{selecting pathways whose PDS is higher than the cutoff value, the default value is NULL}
}
\value{
an Incytr object
}
\description{
Based on the multi-model score, add the KPDS (if applicable) and calculate the PDS (the final score).
}
