% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{Cal_scFC}
\alias{Cal_scFC}
\title{Calculate the fold change (log2FoldChange) for the genes in each L-T pathway based on the scRNA-seq data using DESeq2}
\usage{
Cal_scFC(
  object,
  count.matrix = NULL,
  pseudocount = TRUE,
  fitType = c("parametric", "local", "mean", "glmGamPoi"),
  sfType = c("ratio", "poscounts", "iterate")
)
}
\arguments{
\item{object}{Incytr object}

\item{count.matrix}{the count matrix used to calculate the fold change}

\item{pseudocount}{add 1 pseudocount to avoid division over zero issue. Default is TRUE.}

\item{fitType}{an option from the DESeq2, in default, fitType = c("parametric", "local", "mean", "glmGamPoi")}

\item{sfType}{an option from the DESeq2, in default, sfType = c("ratio", "poscounts", "iterate")}
}
\value{
an Incytr object
}
\description{
Calculate the fold change (log2FoldChange) for the genes in each L-T pathway based on the scRNA-seq data using DESeq2
}
