% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{Cal_SigProb}
\alias{Cal_SigProb}
\title{Calculate the signaling probability and its fold change for each L-T pathway based on the gene expression data}
\usage{
Cal_SigProb(
  object,
  K = 0.5,
  N = 2,
  cutoff_SigProb = NULL,
  correction = 1e-04,
  q = NULL
)
}
\arguments{
\item{object}{Incytr object}

\item{K}{parameter for the half occupation level in the hill function, the default value is 0.5}

\item{N}{Hill coefficient, the defaulty value is 2}

\item{cutoff_SigProb}{cutoff value for selecting the pathways that pass this value in at least one condition. In default, it is NULL.}

\item{correction}{the value added to all the data to avoid division over zero, the default value is 0.0001}

\item{q}{the percentile used to determine "low expression level" and is used when calculating the adjlog2FC, the default value is 0.75}
}
\value{
an Incytr object
}
\description{
Calculate the signaling probability and its fold change for each L-T pathway based on the gene expression data
}
