% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{Cal_SigProb_ptest}
\alias{Cal_SigProb_ptest}
\title{Calculating the signaling probability each L-T pathway using the scRNA-seq data during the permutation test}
\usage{
Cal_SigProb_ptest(object, K = 0.5, N = 2)
}
\arguments{
\item{object}{Incytr object}

\item{K}{parameter for the half occupation level in the hill function, the default value is 0.5 (used when predicting the signaling probability)}

\item{N}{Hill coefficient, the defaulty value is 2 (used when predicting the signaling probability)}
}
\value{
a Incytr object
}
\description{
Calculating the signaling probability each L-T pathway using the scRNA-seq data during the permutation test
}
