% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Incytr_class.R
\name{create_Incytr}
\alias{create_Incytr}
\title{Create a new Incytr object from a data matrix}
\usage{
create_Incytr(
  object,
  meta = NULL,
  sender = NULL,
  receiver = NULL,
  group.by = NULL,
  conditions = NULL,
  assay = NULL,
  do.sparse = T
)
}
\arguments{
\item{object}{a normalized (NOT count) data matrix (genes by cells)}

\item{meta}{a data frame (rows are cells with rownames) consisting of the cell cluster and condition labels}

\item{sender}{a cluster name used as the sender cell group}

\item{receiver}{a cluster name used as the receiver cell group}

\item{group.by}{a char name of the variable in meta data, defining cell groups.
If input is a data matrix and group.by is NULL, the input \code{meta} should contain a column named 'labels',
If input is a Seurat or SingleCellExperiment object, USER must provide \code{group.by} to define the cell groups. e.g, group.by = "ident" for Seurat object}

\item{conditions}{condition labels (no more than two) to use in the analysis.
If the input meta does not have the 'condition' variable, it will be created (as a factor) with one condition label 'condition1',
Then if the input conditions is NULL, the factor levels of the variable 'condition' will be used.}

\item{assay}{Assay to use when the input is a Seurat object}

\item{do.sparse}{whether use sparse format}
}
\value{
an Incytr object
}
\description{
Create a new Incytr object from a data matrix
}
