% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{dataprepare_Expr_bygroup}
\alias{dataprepare_Expr_bygroup}
\title{Calculate the average expression level of each cluster in two conditions from the single-cell data}
\usage{
dataprepare_Expr_bygroup(
  object,
  meta = NULL,
  group.by = NULL,
  conditions = NULL,
  omics = NULL,
  mean_method = NULL
)
}
\arguments{
\item{object}{Incytr object}

\item{meta}{data frame storing the information associated with each cell}

\item{group.by}{variable name used to label the cells}

\item{conditions}{names of the conditions, or levels of the condition factor}

\item{omics}{a char to note the data type, only used in the variable names. Must be NULL or one of following: pr, ps, py.}

\item{mean_method}{the method name used to calculate the average expressed value. NULL is the default value, and the arithmetic mean is used if it is "mean".}
}
\value{
a data frame
}
\description{
Calculate the average expression level of each cluster in two conditions from the single-cell data
}
