% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{Export_results}
\alias{Export_results}
\title{Export the results as a table from the Incytr object}
\usage{
Export_results(object, indicator = FALSE)
}
\arguments{
\item{object}{Incytr object}

\item{indicator}{determine if we need the indicators to show the number of changed components in the pathways. In default, indicator = FALSE}
}
\value{
a data frame
}
\description{
Export the results as a table from the Incytr object
}
