% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{Expr_bygroup}
\alias{Expr_bygroup}
\title{Calculate the average expression level in each cell group for both conditions}
\usage{
Expr_bygroup(object, mean_method = NULL)
}
\arguments{
\item{object}{Incytr object}

\item{mean_method}{the method name used to calculate the average expressed value. NULL is the default value, and the arithmetic mean is used if it is "mean".}
}
\value{
a data frame
}
\description{
Calculate the average expression level in each cell group for both conditions
}
