% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{Find_highexp_gene}
\alias{Find_highexp_gene}
\title{Infer the highly expressed genes of a select cluster}
\usage{
Find_highexp_gene(
  object,
  meta,
  group.by,
  group.select,
  geneuse = NULL,
  cutoff_percentile = 0.5,
  mean_method = NULL
)
}
\arguments{
\item{object}{Incytr object}

\item{meta}{data frame storing the information associated with each cell}

\item{group.by}{variable name used to label the cells}

\item{group.select}{cluster name for inferring highly expressed genes}

\item{geneuse}{a list of genes used for the inference, use NULL if all genes are using. The default value is NULL.}

\item{cutoff_percentile}{the persentile used to identifiy the highly expressed genes. The default value is cutoff_percentile = 0.5, which means if a gene's average expression level is higher than the 50th persentile then it is identified as a highly expressed gene.}

\item{mean_method}{the method name used to calculate the average expressed value. NULL is the default value, and the arithmetic mean is used if it is "mean".}
}
\value{
a data frame
}
\description{
Infer the highly expressed genes of a select cluster
}
