% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{Generate_indicator}
\alias{Generate_indicator}
\title{Recording how many components (in one pathway) is up/down-regulated}
\usage{
Generate_indicator(df, cutoff = NULL)
}
\arguments{
\item{df}{data frame of the L-T pathway information, usually is selected from the Incytr object}

\item{cutoff}{the value defining how much change is up/down-regulated, the default value is 0}
}
\value{
a data frame
}
\description{
Recording how many components (in one pathway) is up/down-regulated
}
