% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Incytr_class.R
\docType{class}
\name{Incytr-class}
\alias{Incytr-class}
\alias{object}
\title{The key slots used in the Incytr object are described below.}
\description{
The key slots used in the Incytr object are described below.
}
\section{Slots}{

\describe{
\item{\code{data.raw}}{raw count data matrix}

\item{\code{data}}{normalized data matrix for CellChat analysis (Genes should be in rows and cells in columns)}

\item{\code{expr.bygroup}}{a list of data frames storing the average gene expression level in each cluster in each condition}

\item{\code{pathways}}{data frame storing the inferred L-R-EM-T pathways}

\item{\code{pathways_5steps}}{data frame storing the inferred 5-step pathways for future exploration}

\item{\code{meta}}{data frame storing the information associated with each cell}

\item{\code{sender}}{a char for the cluster name used as the sender cell group}

\item{\code{receiver}}{a char for the cluster name used as the receiver cell group}

\item{\code{idents}}{factor of the cell cluster labels, with cell barcodes}

\item{\code{conditions}}{a char for the condition labels (no more than two) to use in the analysis}

\item{\code{SigProb}}{data frame storing the predicted signaling probability of each L-R-EM-T pathway in each condition}

\item{\code{p_value}}{data frame storing the p-value of each L-R-EM-T pathway in each condition}

\item{\code{sc_FC}}{list of data frames storing the gene log2FoldChange values calculated using DESeq2}

\item{\code{pr_FC}}{data frame storing the fold change values of each cluster between conditions, calculated via the proteomics data}

\item{\code{ps_FC}}{data frame storing the fold change values of each cluster between conditions, calculated via the ps data}

\item{\code{py_FC}}{data frame storing the fold change values of each cluster between conditions, calculated via the py data}

\item{\code{Evaluation}}{data frame storing the evaluation scores of the L-R-EM-T pathways}

\item{\code{kl}}{data frame of the kinase library}

\item{\code{kl.pathways}}{data frame storing all the inferred signaling-involved kinases (SiKs) of each pathways}

\item{\code{EI}}{list storing the gene excluiveness index (EI) in each condition}

\item{\code{kl.explore}}{data frame storing the inferred singaling-related kinases (SrKs) and their EI}

\item{\code{options}}{List of miscellaneous data, such as parameters used throughout analysis}
}}

