% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{Infer_5step_pathways}
\alias{Infer_5step_pathways}
\title{Infer the 5-step pathways based on the L-T pathway table for fututr exploration}
\usage{
Infer_5step_pathways(object)
}
\arguments{
\item{object}{Incytr object}
}
\value{
an Incytr object
}
\description{
Infer the 5-step pathways based on the L-T pathway table for fututr exploration
}
