% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{Integr_multiomics}
\alias{Integr_multiomics}
\title{Integrate the proteomics and phosphorylation data to the Incytr object, and calculate the log2FC, adjlog2FC}
\usage{
Integr_multiomics(
  object,
  pr.data_condition1 = NULL,
  pr.data_condition2 = NULL,
  pr.correction = NULL,
  pr.q = NULL,
  ps.data_condition1 = NULL,
  ps.data_condition2 = NULL,
  ps.correction = NULL,
  ps.q = NULL,
  py.data_condition1 = NULL,
  py.data_condition2 = NULL,
  py.correction = NULL,
  py.q = NULL
)
}
\arguments{
\item{object}{Incytr object}

\item{pr.data_condition1}{data frame of the proteomics data in condition1, must have the variable "gene_symbol".}

\item{pr.data_condition2}{data frame of the proteomics data in condition2, must have the variable "gene_symbol"}

\item{pr.correction}{the correction used when calculating the fold change to avoid division over zero issue, the default value is NULL}

\item{pr.q}{the percentile used to determine "low expression level" and is used when calculating the adjlog2FC, the default value is 0.75}

\item{ps.data_condition1}{similar to the above}

\item{ps.data_condition2}{similar to the above}

\item{ps.correction}{similar to the above}

\item{ps.q}{similar to the above}

\item{py.data_condition1}{similar to the above}

\item{py.data_condition2}{similar to the above}

\item{py.correction}{similar to the above}

\item{py.q}{similar to the above}
}
\value{
an Incytr object
}
\description{
Integrate the proteomics and phosphorylation data to the Incytr object, and calculate the log2FC, adjlog2FC
}
