% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{logi}
\alias{logi}
\title{Define Logistic function}
\usage{
logi(x, k)
}
\arguments{
\item{x}{input variable}

\item{k}{parameter controls the overall slope}
}
\value{
a value
}
\description{
Define Logistic function
}
