% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{object_update}
\alias{object_update}
\title{Update the Incytr object when the inferred L-T paths have been changed/updated}
\usage{
object_update(object, Path)
}
\arguments{
\item{object}{Incytr object}

\item{Path}{list of the new path names}
}
\value{
an Incytr object
}
\description{
Update the Incytr object when the inferred L-T paths have been changed/updated
}
