% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{Pathway_evaluation}
\alias{Pathway_evaluation}
\title{Calculate the TPDS, PPDS, PhPDS, and multimodel score for the L-T pathways}
\usage{
Pathway_evaluation(
  object,
  score.weight = NULL,
  k_logi = NULL,
  style = NULL,
  abs.value = "None",
  cutoff_TPDS = NULL
)
}
\arguments{
\item{object}{Incytr object}

\item{score.weight}{weight for proteomics, ps, and pr data, the default values are all 0.5.}

\item{k_logi}{parameter for the Logistic funtion, the defaulte value is 2}

\item{style}{the fold change value used in the evaluation, should be either "log2FC" or "aFC", the default value is "aFC"}

\item{abs.value}{the values can be "None", "All", "Ligand", "Receptor", "EM", "Target". The default is "None" means the fold change values used in the evaluation can be positive or negative. If it is "All" then the absolute value of the fold changes will be used, ignoring the direction of the change.}

\item{cutoff_TPDS}{selecting the L-T pathways whose TPDS is higher than the cutoff}
}
\value{
an Incytr object
}
\description{
Calculate the TPDS, PPDS, PhPDS, and multimodel score for the L-T pathways
}
