% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{pathway_inference}
\alias{pathway_inference}
\title{Inferring signaling pathways based on the user-select genes & database}
\usage{
pathway_inference(
  object,
  DB,
  gene.use_Sender = NULL,
  gene.use_Receiver = NULL,
  ligand = NULL,
  receptor = NULL,
  em = NULL,
  target = NULL
)
}
\arguments{
\item{object}{Incytr object}

\item{DB}{database used for the inference}

\item{gene.use_Sender}{a list of gene used as the sender genes (ligands). Used all the genes when it is NULL.}

\item{gene.use_Receiver}{a list of gene used as the receiver genes (receptors, EMs, and targets). Used all the genes when it is NULL.}

\item{ligand}{a list of gene used to filter the ligand}

\item{receptor}{a list of gene used to filter the receptor}

\item{em}{a list of gene used to filter the em}

\item{target}{a list of gene used to filter the target}
}
\value{
an Incytr object
}
\description{
Inferring signaling pathways based on the user-select genes & database
}
