% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{Permutation_test}
\alias{Permutation_test}
\title{Calculate the p_value based on the scRNA-seq data for pathways in each condition using permutation test
an Incytr object}
\usage{
Permutation_test(
  object,
  K = 0.5,
  N = 2,
  nboot = 10,
  seed.use = 1L,
  mean_method = NULL,
  FDR_method = "BH",
  cutoff_p_value = NULL
)
}
\arguments{
\item{object}{Incytr object}

\item{K}{parameter for the half occupation level in the hill function, the default value is 0.5 (used when predicting the signaling probability)}

\item{N}{Hill coefficient, the defaulty value is 2 (used when predicting the signaling probability)}

\item{nboot}{number of permutation test performed, the default value is 10 for a faster running time. However, the reasonable value should be at least 100.}

\item{seed.use}{the seed use in the random selection}

\item{mean_method}{the method name used to calculate the average expressed value. NULL is the default value, and the arithmetic mean is used if it is "mean".}

\item{FDR_method}{the method used in the FDR correction (in the function p.adjust), the default selection is "BH"}

\item{cutoff_p_value}{the cutoff value used to select the pathway that has a lower p-value in at least one condition}
}
\value{
an Incytr object
}
\description{
Calculate the p_value based on the scRNA-seq data for pathways in each condition using permutation test
an Incytr object
}
