% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{proteomics_gene}
\alias{proteomics_gene}
\title{Infer genes with high proteomics fold change}
\usage{
proteomics_gene(
  pr.data_condition1,
  pr.data_condition2,
  cell_group,
  style = "aFC",
  cutoff = 1,
  pr.correction = 1e-05,
  pr.q = NULL
)
}
\arguments{
\item{pr.data_condition1}{data frame for proteomics data in condition 1, must have variable "gene_symbol"}

\item{pr.data_condition2}{data frame for proteomics data in condition 2, must have variable "gene_symbol"}

\item{cell_group}{levels for the cell cluster labels}

\item{style}{select the fold change formula between "log2FC" and "aFC", the default value is "aFC"}

\item{cutoff}{keep the gene with fold change value passing the cutoff, in default cutoff = 1}

\item{pr.correction}{the correction used when calculating the fold change to avoid division over zero issue, the default value is NULL}

\item{pr.q}{the percentile used to determine "low expression level" and is used when calculating the adjlog2FC, the default value is 0.75}
}
\value{
a data frame
}
\description{
Infer genes with high proteomics fold change
}
