% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{Cal_scFC}
\alias{Cal_scFC}
\title{Calculate the fold change (log2FoldChange) for the genes in each L-T pathway based on the expression level calculated by Incytr}
\usage{
Cal_scFC(object, style = "log2FC", correction = 1e-05, q = NULL)
}
\arguments{
\item{object}{Incytr object}

\item{style}{select the fold change formula between "log2FC" and "aFC", the default value is "log2FC"}

\item{correction}{the value added to all the data to avoid division over zero, the default value is 0.0001}

\item{q}{the percentile used to determine "low expression level" and is used when calculating the adjlog2FC, the default value is 0.75}
}
\value{
an Incytr object
}
\description{
Calculate the fold change (log2FoldChange) for the genes in each L-T pathway based on the expression level calculated by Incytr
}
